whatsAppBot-API/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# CORS
starlette>=0.27.0

# Date/Time handling
pytz>=2023.3

# Data processing and analysis
numpy>=1.26.0
pandas>=2.1.1

# Testing
pytest>=7.4.3
httpx>=0.25.0
pytest-asyncio>=0.21.1

# Development tools
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0 


	


 MALE = 17
FEMALE = 13
TOTAL 30

Male Percentage = 17/30 * 360 = 56.7
Female percentage = 13/30 * 360 = 43.3